requirements.txt
alembic==1.7.7
blinker==1.4
Bootstrap-Flask==2.0.2
certifi==2022.5.18.1
cffi==1.15.0
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
dnspython==2.2.1
elastic-transport==8.1.2
elasticsearch==8.2.0
email-validator==1.2.1
Flask==2.1.2
Flask-DebugToolbar==0.13.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
